drf-petclinic-microservices/pyproject.toml
[tool.poetry]
name = "drf-petclinic-microservices"
version = "0.1.0"
description = ""
authors = ["David Pequegnot <david.pequegnot@worldline.com>"]

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.0.1"
djangorestframework = "^3.13.1"
PyYAML = "^6.0"
djangorestframework-camel-case = "^1.3.0"
py-eureka-client = "^0.10.2"
opentelemetry-distro = {extras = ["otlp"], version = "^0.27b0"}
opentelemetry-instrumentation = "^0.27b0"
opentelemetry-instrumentation-django = "^0.27b0"
opentelemetry-instrumentation-dbapi = "^0.27b0"
opentelemetry-instrumentation-sqlite3 = "^0.27b0"
opentelemetry-instrumentation-wsgi = "^0.27b0"
opentelemetry-instrumentation-requests = "^0.27b0"
opentelemetry-propagator-b3 = "^1.8.0"
gunicorn = "^20.1.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
